backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
argon2-cffi
pyjwt
alembic
structlog
python-multipart
slowapi
pyotp
qrcode[pil]
webauthn
requests
httpx
docker
websockets
pyyaml
